requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.0.7
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.3.0
djangorestframework==3.11.0
drf-action-serializer==1.0.5
haversine==2.2.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.1
model-bakery==1.1.0
munch==2.5.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5

haversine~=2.2.0